Add gate

+++ event-management/app/Http/Controllers/Api/AttendeeController.php
@@ -14,6 +14,10 @@ class AttendeeController extends Controller
     use CanLoadRelationships;
 
     private array $relations = ['user'];
+    public function __construct()
+    {
+        $this->middleware('auth:sanctum')->except(['index', 'show', 'update']);
+    }
 
     public function index(Event $event)
     {
@@ -52,6 +56,7 @@ class AttendeeController extends Controller
      */
     public function destroy(string $event, Attendee $attendee)
     {
+        $this->authorize('delete-attendee', $event, $attendee);
         $attendee->delete();
 
         return response(status: 204);
